Re-organize and rename tests
