+++ test/integration/capybara/observations_integration_test.rb
@@ -3,7 +3,7 @@
 require("test_helper")
 
 # Tests which supplement controller/observations_controller_test.rb
-class ObservationsControllerSupplementalTest < CapybaraIntegrationTestCase
+class ObservationsIntegrationTest < CapybaraIntegrationTestCase
   # Prove that when a user "Tests" the text entered in the Textile Sandbox,
   # MO displays what the entered text looks like.
   def test_post_textile
@@ -179,5 +179,18 @@ class ObservationsControllerSupplementalTest < CapybaraIntegrationTestCase
     assert_no_selector("#content div.alert-warning")
     assert_selector("#title",
                     text: "Observation #{obs.id}: #{obs.name.search_name}")
+  end
+
+  # Tests of show_name_helper module
+  # Prove that all these links appear under "Observations of"
+  def test_links_to_observations_of
+    login
+    # on ShowObservation page
+    visit("/names/#{names(:chlorophyllum_rachodes).id}")
+    assert_text(:obss_of_this_name.l)
+    assert_text(:taxon_obss_other_names.l)
+    assert_text(:obss_of_taxon.l)
+    assert_text(:obss_taxon_proposed.l)
+    assert_text(:obss_name_proposed.l)
   end
 end
