+++ test/integration/capybara/sequences_integration_test.rb
@@ -1,14 +1,9 @@
 # frozen_string_literal: true
 
 require("test_helper")
-# require("capybara_helper")
 
 # Test adding, editing, and deleting a Sequence
-# Uses Capybara; MO SessionFormExtensions don't work on form_with forms
-# Use an integration test because click_mo_link needs an id, and
-# there's no  id (at least by default) for a form_with resource-oriented
-# form field.
-class SequencerTest < CapybaraIntegrationTestCase
+class SequencesIntegrationTest < CapybaraIntegrationTestCase
   def test_sequence
     obs = observations(:detailed_unknown_obs)
     sequence_original_count = Sequence.count
